:bug: fix(assistant): enable public assistant sharing within tenant
- Fix message handler access control to use canAccessAssistant
- Non-owners can now chat with public assistants in same tenant
- Centralized access control logic in shared utility
- Enhanced tenant ID fallback logging for security
- Fixed s3Url type definition
